+++ packages/pluggable-widgets-tools/bin/mx-scripts.js
@@ -1,7 +1,9 @@
 #! /usr/bin/env node
-const { spawnSync } = require("child_process");
+const { execSync, spawnSync } = require("child_process");
 const { existsSync } = require("fs");
 const { delimiter, dirname, join, parse } = require("path");
+
+checkNodeVersion();
 
 const [, currentScriptPath, cmd, ...args] = process.argv;
 const toolsRoot = currentScriptPath.endsWith("pluggable-widgets-tools")
@@ -102,4 +104,27 @@ function findNodeModulesBin() {
         parentDir = join(parentDir, "..");
     }
     throw new Error("Cannot find bin folder");
+}
+
+function checkNodeVersion() {
+    console.log("Checking node and npm version...");
+    try {
+        const nodeVersion = extractMajorVersion(execSync("node --version").toString().trim());
+        const npmVersion = extractMajorVersion(execSync("npm --version").toString().trim());
+        if (nodeVersion < 16) {
+            console.error("To build this widget a minimum node version 16.0.0 is required. Please upgrade your node version!");
+            process.exit(1);
+        }
+        if (npmVersion < 8) {
+            console.error("To build this widget a minimum npm version 8.0.0 is required. Please upgrade your npm version!");
+            process.exit(1);
+        }
+    } catch(e) {
+        throw new Error("Cannot find node or npm to determine the version")
+    }
+}
+
+function extractMajorVersion(version) {
+    const majorVersion = version.replace(/^\D+/, '').split(".")[0];
+    return Number(majorVersion);
 }