+++ packages/pluggable-widgets-tools/configs/hot-typescript.ts
@@ -1,0 +1,3 @@
+export function hot<T>(widget: T): T {
+    return widget;
+}