+++ packages/pluggable-widgets-tools/test-config/jest.config.js
@@ -4,28 +4,31 @@ const projectDir = process.cwd();
 
 module.exports = {
     clearMocks: true,
+    testRunner: "jest-jasmine2",
     rootDir: join(projectDir, "src"),
     setupFilesAfterEnv: [join(__dirname, "test-index.js")],
     snapshotSerializers: ["enzyme-to-json/serializer"],
     testMatch: ["<rootDir>/**/*.spec.{js,jsx,ts,tsx}"],
     transform: {
-        "\\.tsx?$": [
+        "^.+\\.tsx?$": [
             "ts-jest",
             {
                 isolatedModules: true,
                 tsconfig: { module: "commonjs" }
             }
         ],
-        "\\.jsx?$": join(__dirname, "transform.js"),
-        "^.+\\.svg$": "jest-svg-transformer"
+        "^.+\\.jsx?$": join(__dirname, "transform.js"),
+        "^.+\\.svg$": join(__dirname, "jest-svg-transformer")
     },
     moduleNameMapper: {
         "\\.(css|less|scss|sass)$": "identity-obj-proxy",
         "mendix/components/web/Icon": join(__dirname, "__mocks__/WebIcon"),
         "mendix/filters/builders": join(__dirname, "__mocks__/FilterBuilders"),
-        "\\.png$": join(__dirname, "assetsTransformer.js")
+        "\\.png$": join(__dirname, "assetsTransformer.js"),
+        "react-hot-loader/root": join(__dirname, "__mocks__/hot")
     },
+    moduleDirectories: ["node_modules", join(projectDir, "node_modules")],
     collectCoverage: !process.env.CI,
-    coverageDirectory: "<rootDir>/../dist/coverage",
+    coverageDirectory: join(projectDir, "dist/coverage"),
     testEnvironment: "jsdom"
 };