+++ packages/pluggable-widgets-tools/test-config/jest.native.config.js
@@ -3,33 +3,41 @@ const { join } = require("path");
 const projectDir = process.cwd();
 
 module.exports = {
-    preset: "react-native",
+    preset: hasDependency("@testing-library/react-native") ? "@testing-library/react-native" : "react-native",
+    testRunner: "jest-jasmine2",
     clearMocks: true,
+    haste: {
+        defaultPlatform: "android",
+        platforms: ["android", "ios", "native"],
+    },
     rootDir: join(projectDir, "src"),
     setupFilesAfterEnv: [
         join(__dirname, "test-index-native.js"),
-        ...(hasDependency("react-native-gesture-handler") ? ["react-native-gesture-handler/jestSetup.js"] : [])
+        ...(hasDependency("react-native-gesture-handler") ? ["react-native-gesture-handler/jestSetup.js"] : []),
+        ...(hasDependency("@testing-library/jest-native") ? ["@testing-library/jest-native/extend-expect"] : [])
     ],
     snapshotSerializers: ["enzyme-to-json/serializer"],
     testMatch: ["<rootDir>/**/*.spec.{js,jsx,ts,tsx}"],
-    transformIgnorePatterns: ["node_modules/(?!.*react-native)(?!victory-)"],
+    transformIgnorePatterns: ["node_modules/(?!(.*react-native.*|victory-)/)"],
     transform: {
-        "\\.tsx?$": [
+        "^.+\\.tsx?$": [
             "ts-jest",
             {
                 isolatedModules: true,
                 tsconfig: { module: "commonjs" }
             }
         ],
-        "\\.jsx?$": join(__dirname, "transform-native.js")
+        "^.+\\.jsx?$": join(__dirname, "transform-native.js")
     },
     moduleNameMapper: {
         "mendix/components/native/Icon": join(__dirname, "__mocks__/NativeIcon"),
         "mendix/components/native/Image": join(__dirname, "__mocks__/NativeImage"),
-        "mendix/filters/builders": join(__dirname, "__mocks__/FilterBuilders")
+        "mendix/filters/builders": join(__dirname, "__mocks__/FilterBuilders"),
+        "react-hot-loader/root": join(__dirname, "__mocks__/hot")
     },
+    moduleDirectories: ["node_modules", join(projectDir, "node_modules")],
     collectCoverage: !process.env.CI,
-    coverageDirectory: "<rootDir>/../dist/coverage",
+    coverageDirectory: join(projectDir, "dist/coverage"),
     testEnvironment: "jsdom"
 };