+++ packages/pluggable-widgets-tools/test-config/transform.js
@@ -1,8 +1,11 @@
 module.exports = require("babel-jest").createTransformer({
-    presets: ["@babel/preset-env", "@babel/preset-react"],
+    presets: [
+        ["@babel/preset-env", { "modules": "auto" }],
+        "@babel/preset-react"
+    ],
     plugins: [
-        ["@babel/plugin-proposal-class-properties", { loose: true }],
-        ["@babel/plugin-proposal-private-methods", { loose: true }],
+        "@babel/plugin-proposal-class-properties",
+        "@babel/plugin-proposal-private-methods",
         ["@babel/plugin-transform-react-jsx", { pragma: "createElement" }]
     ]
 });